requirements.txt
annotated-types==0.7.0
cli_helpers==2.3.1
click==8.1.7
configobj==5.0.8
markdown-it-py==3.0.0
mdurl==0.1.2
pgcli==4.1.0
pgspecial==2.1.2
prompt_toolkit==3.0.47
psycopg==3.1.17
psycopg-binary==3.2.1
psycopg-pool==3.2.2
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-dotenv==1.0.1
rich==13.7.1
setproctitle==1.3.3
six==1.16.0
sqlparse==0.5.1
tabulate==0.9.0
typing_extensions==4.12.2
wcwidth==0.2.13
